Create Apple and Orange
Sam's house has an apple tree and an orange tree that yield an abundance of fruit.
his house, where s is the start point, and t  is the endpoint. The apple tree is to the left of his house, and the orange tree is to its right. You can assume the trees are located on a single point, where the apple tree is at point a, and the orange tree is at point b.
When fruit falls from its tree, it lands d units of distance from its tree of origin along the x-axis. A negative value of means the fruit fell d units to the tree's left, and a positive value of means it falls d units to the tree's right.

Given the value of  for  apples and  oranges, determine how many apples and oranges will fall on Sam's house